pyproject.toml
[tool.poetry]
name = "pdf-translator"
version = "0.1.0"
description = ""
authors = ["Yuki Oshima <39944763+os1ma@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^0.0.250"
openai = "^0.27.8"
python-dotenv = "^1.0.0"
pypdf = "^3.14.0"
click = "^8.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
